feat(typeahead) stop selecting first result when displaying matches
Switches from always setting the first result to be the active one, and
instead initially displays the result list with no active item.
After the pressing down/up keys the first/last item will be selected.

Also cleans up the keyboard interaction when alternating between
navigating between matches with arrow keys and typing additional
characters, by keeping the previous match selected if it continues to be
a match with the new results, instead of the previous behavior of
resetting to the first item selected.

Closes #3965