fix: XSS issue in typeahead component
Even though it is possible to expect that the provided options are safe,
such assumption will cause weird highlighting behavior. For example, if
`<foo>` is provided as option, the safe version would be `&lt;foo&gt;`.
This means that user will see the `<` & `>` characters but if they
search for it, nothing would match.

Hence, this is a better to escape the string in typeahead component so
the search and highlight functionality will remain properly functional.